refactor: add CompetitionSection
