+++ apps/animated-landing/src/components/CompetitionCard.tsx
@@ -9,10 +9,10 @@ interface CompetitionCardProps {
 
 export default function CompetitionCard({ src, title, children }: CompetitionCardProps) {
   return (
-    <article className="bg-card-background rounded-card w-full p-0.5">
-      <div className="bg-n3 rounded-card flex flex-col items-center px-6 py-8">
+    <article className="bg-card-background rounded-card max-w-[485px] p-0.5 md:w-full">
+      <div className="bg-n3 rounded-card section h-full px-6 py-8 md:p-10">
         <Image alt={title} src={src} />
-        <h3 className="text-p1 text-ch-h3 py-2 font-bold [text-shadow:theme(boxShadow.purple)]">
+        <h3 className="text-p1 text-ch-h3 md:text-ch-h2 py-2 font-bold [text-shadow:theme(boxShadow.purple)]">
           {title}
         </h3>
         {children}
